pyproject.toml
[project]
name = "zzz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.1.0",
    "fastapi>=0.117.1",
    "playwright>=1.55.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.36.0",
]
